face_module/requirements.txt
face_recognition==1.3.0
dlib==19.24.2
numpy==1.24.3
Pillow==10.0.1
opencv-python==4.8.1.78
